requirements.txt
# Core dependencies
python-dotenv==1.2.2
aiosqlite==0.22.1
pytest==9.0.2
pytest-asyncio==0.26.0

# Content extraction
youtube-transcript-api==1.2.4
beautifulsoup4==4.14.3
requests==2.32.5
PyMuPDF==1.27.2.2

# CLI prompts
questionary==2.1.1

# Web research
tavily-python==0.5.0

# YouTube search
yt-dlp>=2024.0.0

# Optional (for future phases)
# sentence-transformers  # For embeddings
# tiktoken  # For token counting
